add metrics metrics manager component
para sacarle la responsabilidad de recopilar métricas de cada episodio y de cada algoritmo a la simulación

también refactoricé un toke el main, y se le sacaron todos los parámetros de las métricas a la simulación
